Changed volume size variables from int to unsigned long int

+++ lib/qhg_ape_smear.c
@@ -17,9 +17,9 @@ qhg_ape_smear_3d_iter(qhg_gauge_field out, qhg_gauge_field in, double alpha_in)
   {
 #endif
     
-  int vol = in.lat->vol;
-  int lvol = in.lat->lvol;
-  int **nn = in.lat->nn;
+  unsigned long int vol = in.lat->vol;
+  unsigned long int lvol = in.lat->lvol;
+  unsigned long int **nn = in.lat->nn;
   double alpha = 4.0*alpha_in / (1.0 + 4.0*alpha_in);
 
 #ifdef QHG_OMP
@@ -46,15 +46,15 @@ qhg_ape_smear_3d_iter(qhg_gauge_field out, qhg_gauge_field in, double alpha_in)
 #ifdef QHG_OMP
 #pragma omp for
 #endif
-    for(int v00=0; v00<lvol; v00++) {
+    for(unsigned long int v00=0; v00<lvol; v00++) {
       su3_linalg_zero(staple);
 
       /* create the staple term for this mu direction */
       for(int nu=(mu==1?2:1); nu<ND; nu==mu-1?nu+=2:nu++)	{
-      	int vp0 = nn[mu][v00];
-      	int v0p = nn[nu][v00];
-      	int v0m = nn[nu+ND][v00];	
-      	int vpm = nn[nu+ND][nn[mu][v00]];
+      	unsigned long int vp0 = nn[mu][v00];
+      	unsigned long int v0p = nn[nu][v00];
+      	unsigned long int v0m = nn[nu+ND][v00];	
+      	unsigned long int vpm = nn[nu+ND][nn[mu][v00]];
 	
       	/* Fwd staple */
       	u0 = &U[G(v00, nu)];
